+++ blacksheep/plugins/__init__.py
@@ -1,0 +1,7 @@
+from blacksheep.plugins.json import JSONPlugin
+
+json = JSONPlugin()
+
+__all__ = [
+    "json",
+]