+++ blacksheep/server/responses.py
@@ -4,11 +4,9 @@ from functools import lru_cache
 from io import BytesIO
 from typing import Any, AnyStr, AsyncIterable, Callable, Union
 
-from essentials.json import FriendlyEncoder
-from essentials.json import dumps as friendly_dumps
-
 from blacksheep import Content, JSONContent, Response, StreamedContent, TextContent
 from blacksheep.common.files.asyncfs import FilesHandler
+from blacksheep.plugins import json as json_plugin
 
 MessageType = Any
 
@@ -27,7 +25,7 @@ def _ensure_bytes(value: AnyStr) -> bytes:
 
 
 def _json_serialize(obj) -> str:
-    return friendly_dumps(obj, cls=FriendlyEncoder, separators=(",", ":"))
+    return json_plugin.dumps(obj)
 
 
 def _json_content(obj) -> JSONContent:
@@ -180,7 +178,7 @@ def html(value: str, status: int = 200) -> Response:
     )
 
 
-def json(data: Any, status: int = 200, dumps=friendly_dumps) -> Response:
+def json(data: Any, status: int = 200) -> Response:
     """
     Returns a response with application/json content,
     and given status (default HTTP 200 OK).
@@ -190,13 +188,15 @@ def json(data: Any, status: int = 200, dumps=friendly_dumps) -> Response:
         None,
         Content(
             b"application/json",
-            dumps(data, cls=FriendlyEncoder, separators=(",", ":")).encode("utf8"),
+            json_plugin.dumps(data).encode("utf8"),
         ),
     )
 
 
 def pretty_json(
-    data: Any, status: int = 200, dumps=friendly_dumps, indent: int = 4
+    data: Any,
+    status: int = 200,
+    indent: int = 4,
 ) -> Response:
     """
     Returns a response with indented application/json content,
@@ -207,7 +207,7 @@ def pretty_json(
         None,
         Content(
             b"application/json",
-            dumps(data, cls=FriendlyEncoder, indent=indent).encode("utf8"),
+            json_plugin.pretty_dumps(data).encode("utf8"),
         ),
     )